requirements.txt
streamlit>=1.32.0
google-genai>=1.71.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0